requirements.txt
PyGithub
pandas
pyyaml
python-dotenv
tabulate
requests
